+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_field_value_factory.dart
@@ -10,7 +10,7 @@ import 'method_channel_field_value.dart';
 
 /// An implementation of [FieldValueFactoryPlatform] that is suitable to be used
 /// on mobile where communication relies on [MethodChannel]
-class MethodChannelFieldValueFactory extends FieldValueFactoryPlatform {
+class MethodChannelFieldValueFactory extends FieldValueFactoryPlatform<MethodChannelFieldValue> {
   @override
   MethodChannelFieldValue arrayRemove(List elements) =>
       MethodChannelFieldValue(FieldValueType.arrayRemove, elements);