+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/utils/firestore_message_codec.dart
@@ -79,6 +79,8 @@ class FirestoreMessageCodec extends StandardMessageCodec {
       final int code = _kFieldPathCodes[value.type];
       assert(code != null);
       buffer.putUint8(code);
+    } else if (value is MethodChannelFieldValue) {
+      assert(false);
     } else {
       super.writeValue(buffer, value);
     }